# docs/openApi.yaml
openapi: 3.0.0
info:
  title: Sample Payment API
  description: 簡易な支払い管理APIです
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: local server
...
paths:
  /users/signup:
    post:
      tags:
        - users
      summary: ユーザ登録
      description: ユーザの新規登録を行う
      parameters: []
      requestBody:
        description: user to create
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserSignupRequest'
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
  /users/auth:
    post:
      tags:
        - users
      summary: ユーザ認証
      description: ユーザの認証を行う
      parameters: []
      requestBody:
        description: user to create
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserAuthRequest'
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAuthResponse'
...
components:
  securitySchemes:
    TokenAuth:
      type: apiKey
      in: header
      name: x-api-token
  schemas:
    UserSignupRequest:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
    UserAuthRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
        password:
          type: string
...
    UserAuthResponse:
      type: object
      required:
        - token
      properties:
        token:
          type: string
    UserResponse:
      type: object
      required:
        - id
        - name
        - email
        - encryptedPassword
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        encryptedPassword:
          type: string
...